refactor(events): deduplicate R3/Reactive generators via shared shproj
Events was the last domain without a .SourceGenerators.Shared shproj,
duplicating 14 files (~6178 lines) between R3 and Reactive generators.

Introduce Observables.Events.SourceGenerators.Shared shproj with 14
shared .cs files under namespace Observables.Events.Generators. Backend
switching uses #if EVENTS_R3 / #else preprocessor directives for the
~30 backend-specific code paths (type names, namespace constants, using
directives, R3-only EventObservable bridge emission).

Method names unified across backends (R3ObservableType -> ObservableType,
RxFromEventInvocation -> FromEventInvocation, etc.) with a Reactive-only
private helper ObservableFromEventInvocation preserved under #if !EVENTS_R3.

Net reduction: ~3000+ lines of duplicated code. R3/Reactive generator
projects now contain zero .cs files (csproj + Import only), matching the
pattern established by the other 7 domains.
