--- Observables.Events/Observables.Events.Reactive.SourceGenerators/ObservableEvents/ObservableEventsEntryKind.cs
@@ -1,11 +1,0 @@
-namespace Observables.Events.Reactive.SourceGenerators;
-
-internal enum ObservableEventsEntryKind
-{
-    Events,
-    EventHandlers,
-    RoutedEvents,
-    RoutedEventHandlers,
-    AttachedRoutedEvent,
-    AttachedRoutedEventHandler,
-}
