+++ Observables.Events/Observables.Events.SourceGenerators.Shared/ObservableEvents/ObservableEventsEntryKind.cs
@@ -1,4 +1,4 @@
-namespace Observables.Events.R3.SourceGenerators;
+namespace Observables.Events.Generators;
 
 internal enum ObservableEventsEntryKind
 {
