+++ Observables.Events/Observables.Events.SourceGenerators.Shared/Observables.Events.SourceGenerators.Shared.projitems
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <MSBuildAllProjects Condition="'$(MSBuildVersion)' == '' Or '$(MSBuildVersion)' &lt; '16.0'">$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
+    <HasSharedItems>true</HasSharedItems>
+    <SharedGUID>a1b2c3d4-e5f6-7890-abcd-ef1234567890</SharedGUID>
+  </PropertyGroup>
+  <PropertyGroup Label="Configuration">
+    <Import_RootNamespace>Observables.Events.Generators</Import_RootNamespace>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)EventsBootstrapSyntaxFactory.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ObservableEventsSyntaxFactory.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ObservableEventsGenerator.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ObservableEventsGenerator.AttachedRouted.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ObservableEventsGenerator.Discovery.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ObservableEventsGenerator.EventProperties.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ObservableEventsGenerator.GenericConstraints.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ObservableEventsGenerator.Helpers.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ObservableEventsGenerator.InterfaceEmission.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ObservableEventsGenerator.InterfacePipeline.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ObservableEventsGenerator.RoutedDetection.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ObservableEvents\ObservableEventsConstants.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ObservableEvents\ObservableEventsEntryKind.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ObservableEvents\ObservableEventsModels.cs" />
+  </ItemGroup>
+</Project>
