+++ Observables.slnx
@@ -43,6 +43,8 @@
 
     <Project Path="Observables.Events/Observables.Events.Package/Observables.Events.Package.csproj" />
 
+    <Project Path="Observables.Events/Observables.Events.SourceGenerators.Shared/Observables.Events.SourceGenerators.Shared.shproj" Id="a1b2c3d4-e5f6-7890-abcd-ef1234567890" />
+
     <Project Path="Observables.Events/Observables.Events.Reactive.SourceGenerators/Observables.Events.Reactive.SourceGenerators.csproj" />
 
     <Project Path="Observables.Events/Observables.Events.Reactive.SourceGenerators.Tests/Observables.Events.Reactive.SourceGenerators.Tests.csproj" />
